# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Snappfood API",
        "description": "API documentation for the Snappfood.",
        "contact": {
            "email": "farshidboroomand@gmail.com.com"
        },
        "version": "V1"
    },
    "paths": {
...
        "/api/v1/wallets/{wallet_id}/withdrawals/sheba": {
            "post": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Create a withdrawal request to a sheba number",
                "description": "Creates a withdrawal request from the user's wallet to a given sheba number.",
                "operationId": "createShebaWithdrawal",
                "parameters": [
                    {
                        "name": "wallet_id",
                        "in": "path",
                        "description": "UUID of the wallet",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "d4211cc3-c2a2-4eeb-8d4d-bf0c4fc64547"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "price",
                                    "from_sheba_number",
                                    "to_sheba_number"
                                ],
                                "properties": {
                                    "price": {
                                        "type": "integer",
                                        "example": 100000
                                    },
                                    "from_sheba_number": {
                                        "type": "string",
                                        "example": "IR0000314596741760846999"
                                    },
                                    "to_sheba_number": {
                                        "type": "string",
                                        "example": "IR0000314596741760846888"
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Transfer for monthly expenses"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Withdrawal request created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "withdrawalId": {
                                                    "type": "string",
                                                    "example": "cfffe99a-53fc-4abf-87c3-f7c5e58d022a"
                                                },
                                                "amount": {
                                                    "type": "integer",
                                                    "example": 100000
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "note": {
                                                    "type": "string",
                                                    "example": "Transfer for monthly expenses"
                                                },
                                                "fromShebaNumber": {
                                                    "type": "string",
                                                    "example": "IR0000314596741760846999"
                                                },
                                                "toShebaNumber": {
                                                    "type": "string",
                                                    "example": "IR0000314596741760846888"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Client error (e.g., insufficient balance or sheba mismatch)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Insufficient balance"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "price": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The price must be at least 10000."
                                                    }
                                                },
                                                "from_sheba_number": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The from sheba number format is invalid."
                                                    }
                                                },
                                                "to_sheba_number": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The to sheba number format is invalid."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    },
...
}